Thought I would get a messege prompt last commit, lesson learned.
I've for the first time tried to structure my own cmake project. Making libraries and such. Had a lot to learn and still do. But this was a very useful experience and now I have a much better grasp on what cmake is, what it does and how to use it.

Along with that learning to setup my own github repo, something i've actually not done before. So learning to manage that has also been fun. Messed up the gitignore first attempt so thats why so many files changed last commit when i unfucked(sorry) it.